requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
Werkzeug>=3.0
pyyaml>=6.0
pytest>=8.0
